more Mozilla API; milliseconds must be float
git-svn-id: http://ocamljs.googlecode.com/svn/trunk@17 27578800-e353-0410-a23a-a7c9f63c6ccd

+++ src/libs/javascript/javascript.ml
@@ -34,10 +34,10 @@ module Date =
     external _new_noargs : string -> t = "$new"
 
     let new_ () = _new_noargs "Date"
-    external new_milliseconds : int -> t = "$new" "Date"
+    external new_milliseconds : float -> t = "$new" "Date"
     external new_parts : int -> int -> int -> int -> int -> int -> int -> t = "$new" "Date"
 
-    external getTime : t -> int = "#getTime"
+    external getTime : t -> float = "#getTime"
     external getDate : t -> int = "#getDate"
     external getMonth : t -> int = "#getMonth"
     external getHours : t -> int = "#getHours"
